RecipesApp : change calories type from int to float and update all dependencies [add util method]

+++ RecipesApp/RecipesApp.Application/Ingredients/Commands/CreateIngredient.cs
@@ -7,7 +7,7 @@ namespace RecipesApp.Application.Ingredients.Commands
     {
         public string Name { get; set; }
         public IngredientCategory Category { get; set; }
-        public int Calories { get; set; }
+        public float Calories { get; set; }
         public float Fats { get; set; }
         public float Carbs { get; set; }
         public float Proteins { get; set; }
