+++ RecipesApp/RecipesApp.Application/Ingredients/Commands/UpdateIngredient.cs
@@ -8,7 +8,7 @@ namespace RecipesApp.Application.Ingredients.Commands
         public int IngredientId { get; set; }
         public string Name { get; set; }
         public IngredientCategory Category { get; set; }
-        public int Calories { get; set; }
+        public float Calories { get; set; }
         public float Fats { get; set; }
         public float Carbs { get; set; }
         public float Proteins { get; set; }
