+++ RecipesApp/RecipesApp.Application/MealPlannerFeature/Commands/GenerateMealPlan.cs
@@ -7,7 +7,7 @@ namespace RecipesApp.Application.MealPlannerFeature.Commands
     public class GenerateMealPlan : IRequest<MealPlan>
     {
         public MealType MealType { get; set; }
-        public int Calories { get; set; }
+        public float Calories { get; set; }
         public List<Recipe> Recipes { get; set; }
     }
 }
